pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retiroflow"
version = "0.1.0"
description = "AI-powered wellness retreat management & booking SaaS for Oaxaca, Mexico"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "RetiroFlow"}]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "anthropic>=0.40.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "duckduckgo-search>=6.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
]

[project.scripts]
retiroflow = "retiroflow.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
